[Search] Fix Add Inference Endpoint API call (#210243)
## Summary

This PR includes
- Remove `task_settings` from API call
- Remove `Service_providers` definition from the plugin to use the
global definition.

### Checklist

Check the PR satisfies following conditions.

Reviewers should verify this PR satisfies this list as well.

- [X] Any text added follows [EUI's writing
guidelines](https://elastic.github.io/eui/#/guidelines/writing), uses
sentence case text and includes [i18n
support](https://github.com/elastic/kibana/blob/main/src/platform/packages/shared/kbn-i18n/README.md)
- [X] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

---------

(cherry picked from commit 2e29a504a749c5ce4bd2f167e38d13e4a7952485)
