+++ x-pack/platform/packages/shared/ml/trained_models_utils/src/constants/trained_models.ts
@@ -223,7 +223,7 @@ export interface GetModelDownloadConfigOptions {
 }
 
 export interface LocalInferenceServiceSettings {
-  service: 'elser' | 'elasticsearch';
+  service: 'elasticsearch';
   service_settings: {
     num_allocations: number;
     num_threads: number;
