+++ x-pack/solutions/search/plugins/search_inference_endpoints/public/components/all_inference_endpoints/filter/service_provider_filter.tsx
@@ -6,8 +6,8 @@
  */
 
 import React from 'react';
-import { SERVICE_PROVIDERS } from '../render_table_columns/render_service_provider/service_provider';
-import type { FilterOptions, ServiceProviderKeys } from '../types';
+import { SERVICE_PROVIDERS, ServiceProviderKeys } from '@kbn/inference-endpoint-ui-common';
+import type { FilterOptions } from '../types';
 import { MultiSelectFilter, MultiSelectFilterOption } from './multi_select_filter';
 import * as i18n from './translations';
